feat(core): convert the base classes to TypeScript (spec 06, Stage 4)
Convert the type-foundation layer: Particle, Behaviour, Initializer,
Zone, BaseRenderer, plus the root and core constants.

- Cross-references between the base classes use `import type` (Particle
  <-> Behaviour resolve here; Emitter/System are forward-refs that
  sharpen when those convert) — no `any`.
- Particle's fields are fully declared (Vector3D positions, Behaviour[],
  EasingFunction, a typed `old`/`color`, `Emitter | null` parent);
  `body` is `unknown` (genuinely polymorphic).
- Relaxed Util.setPrototypeByObj's constraint to `T extends object` so
  class instances qualify.
- Add @types/node + tsconfig `types: ["node"]` for the `process`
  reference in `__DEV__` (three stays import-resolved, unaffected).
- Tightened math/integration's `particle` from the Stage-2 forward-ref
  `any` to the real `Particle` type — closes the last `any` in src
  (only ArraySpan's intentional `Span<any>` default remains).

Types-only, no runtime change. Verified: typecheck, lint (0 errors),
245 tests, build (bundle unchanged).
