+++ package.json
@@ -63,6 +63,7 @@
   },
   "devDependencies": {
     "@playwright/test": "^1.61.1",
+    "@types/node": "^26.1.2",
     "@types/three": "^0.185.1",
     "@vitest/coverage-v8": "^4.1.10",
     "chai": "^4.2.0",
