+++ tsconfig.json
@@ -4,6 +4,7 @@
     "module": "esnext",
     "moduleResolution": "bundler",
     "lib": ["es2020", "dom"],
+    "types": ["node"],
     "allowJs": true,
     "checkJs": false,
     "strict": false,
